% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{BinaryEncoder}
\alias{BinaryEncoder}
\title{Binary Encoder}
\usage{
BinaryEncoder(model_dict, ranked_data)
}
\arguments{
\item{model_dict}{The model dictionary containing layer_dictionaries of weights and parameters.}

\item{ranked_data}{The input ranked data.}
}
\value{
A list containing the sample_size X layer_number sized prob_pred matrix of predicted probabilities and
sample_size sized vector of predicted potency score containing values in the range 0-1.
}
\description{
This function generates predicted probabilities and predicted potency score of each sample based on a given model
by iterating over all layers of that model and aggregating results
}
\examples{
encoded_output <- BinaryEncoder(model_dict, ranked_data)
}
