% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{BinaryModule}
\alias{BinaryModule}
\title{Binary Module}
\usage{
BinaryModule(layer_dict, ranked_data, eps = 1e-05)
}
\arguments{
\item{layer_dict}{The layer dictionary containing weights and parameters.}

\item{ranked_data}{The input ranked data.}

\item{eps}{The epsilon value for numerical stability in calculations (default: 1e-05).}
}
\value{
A list containing the calculated matrix R_all before batch normalization and the vector pred of predictions
}
\description{
This function generates predictions for each layer using the input parameter dictionary for a given layer
}
\examples{
output <- BinaryModule(layer_dict, ranked_data)
}
