% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{binData}
\alias{binData}
\title{Binning}
\usage{
binData(predicted_df)
}
\arguments{
\item{predicted_df}{dataframe of 4 columns: predicted potency scores and categories (CytoTRACE 2 model predicted values from predict function)
and smoothed potency score and potency categories (processed values from smoothData function)}
}
\value{
dataframe with updated predicted potency scores and categories
}
\description{
This function performs a binning procedure by ranking cells within their predicted
potency category and arranging cells uniformly by rank per potency category
within equal length partitions of the unit interval, yielding binned smooth potency score.
}
\examples{
predicted_df --> output of smoothData
cytotrace <- binData(predicted_df)
}
