% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotrace2.R
\name{cytotrace2}
\alias{cytotrace2}
\title{cytotrace2: Predicting Cellular Potency Score and Category}
\usage{
cytotrace2(
  input,
  species = "mouse",
  is_seurat = FALSE,
  slot_type = "counts",
  full_model = FALSE,
  batch_size = 10000,
  smooth_batch_size = 1000,
  parallelize_models = TRUE,
  parallelize_smoothing = TRUE,
  ncores = NULL,
  max_pcs = 200,
  seed = 14
)
}
\arguments{
\item{input}{The input single-cell RNA-seq data. It can be either an expression matrix
(rows as genes, columns as cells) or the filepath to .tsv file containing the data table,
or a Seurat object or the filepath to an .rds file containing a Seurat object.}

\item{species}{String, indicates the species name from which the gene names for the
input data come from (two options: "human" or "mouse", default is "mouse").}

\item{is_seurat}{Logical, indicating whether the input is a Seurat object/filepath to a
Seurat object (default is FALSE).}

\item{slot_type}{Character, indicating the type of slot to access from "RNA" assay if provided is a Seurat object &
is_seurat = TRUE. Can be 'counts' or 'data' (default is 'counts')}

\item{full_model}{Logical, indicating whether to predict based on the full ensemble of 17 models
or a reduced ensemble of 5 most predictive models (default is FALSE).}

\item{batch_size}{Integer or NULL, indicating the number of cells to subsample for the pipeline steps.
No subsampling if NULL.(default is 10000, recommended value for input data size > 10K cells).}

\item{smooth_batch_size}{Integer or NULL, indicating the number of cells to subsample further
within the batch_size for the smoothing step of the pipeline. No subsampling if NULL.
(default is 1000, recommended value for input data size > 1K cells).}

\item{parallelize_models}{Logical, indicating whether to run the prediction function on
models in parallel on multiple threads (default is TRUE).}

\item{parallelize_smoothing}{Logical, indicating whether to run the smoothing function
on subsamples in parallel on multiple threads (default is TRUE).}

\item{ncores}{Integer, indicating the number of cores to utilize when parallelize_models
and/or parallelize_smoothing are TRUE (default is NULL, pipeline detects the number of
available cores and runs on that number; for Windows, it will be set to 1).}

\item{max_pcs}{Integer, indicating the maximum number of principal components to use in the smoothing by kNN step (default is 200).}

\item{seed}{Integer, specifying the seed for reproducibility in random processes (default is 14).}
}
\value{
If `is_seurat` is FALSE, the function returns a dataframe containing predicted
raw and postprocessed cellular potency category and potency scores
If `is_seurat` is TRUE, adds metadata elements for all columns of prediction dataframe,
and returns the modified Seurat object.
}
\description{
This function generates single-cell predictions of developmental potential from scRNA-seq data.
It takes in a matrix of gene expression values where columns are cells and rows are genes \cr\cr
}
\details{
The cytotrace2 function performs the following steps:

1. Loads and preprocesses the input data.
2. Predicts single-cell developmental potential (with discrete potency categories and continuous (range 0-1) score).
3. Smoothes the predicted values using diffusion and rescales using a combination of binning and smoothing by kNN approaches.
}
\section{[LICENSE]](https}{
//github.com/digitalcytometry/cytotrace2/blob/master/LICENSE).
}

\examples{
# example run with a scRNA-seq dataset (10x) encompassing 2280 cells from murine pancreatic epithelium (Bastidas-Ponce et al., 2019)
# download the .rds file (this will download the file to your working directory)
download.file("https://drive.google.com/uc?export=download&id=1ivi9TBlmzVTDGzNWQrXXeyL68Wug989K", "Pancreas_10x_downsampled.rds")
# load rds
data <- readRDS("Pancreas_10x_downsampled.rds")
# extract expression data
expression_data <- data$expression_data
# running CytoTRACE 2 main function - cytotrace2 - with default parameters
cytotrace2_result <- cytotrace2(expression_data)

}
\references{
Link to the publication will be added upon being published
}
\seealso{
See https://github.com/digitalcytometry/cytotrace2 for a detailed package description and instructions
}
\author{
Minji Kang, Jose Juan Almagro Armenteros, Gunsagar Gulati, Rachel Gleyzer, Erin Brown and Susanna Avagyan.
}
