% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{disp_fn}
\alias{disp_fn}
\title{Dispersion Function}
\usage{
disp_fn(x)
}
\arguments{
\item{x}{A numeric vector for which the dispersion metric needs to be calculated.}
}
\value{
A numeric value representing the dispersion metric of the input vector.
}
\description{
This function calculates a dispersion metric for a given numeric vector.
}
