% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictData}
\alias{predictData}
\title{Prediction}
\usage{
predictData(parameter_dict, ranked_data, log2_data, parallelize_models, ncores)
}
\arguments{
\item{parameter_dict}{The ensemble parameter dictionary containing model_dictionaries of weights and parameters.}

\item{ranked_data}{The input ranked data.}

\item{log2_data}{The input log2 data.}

\item{parallelize_models}{boolean value, if TRUE models of the ensemble will run on multiple threads}

\item{ncores}{integer indicating the number of cores to utilize}
}
\value{
The dataframe containing predicted potency category and score for each sample (with sample names as rownames) prior to postprocessing
}
\description{
This function generates predicted probabilities and predicted potency score of each sample based on a given
ensemble of models by iterating over all models of the set of models and aggregating results.
It then performs post processing of predicted probabilities and predicted potency scores
to generate 1 categorical label and 1 continuous score value for each sample
by aggregating results over all models and layers.
}
\examples{
parameter_dict <- readRDS(system.file("extdata", "parameter_dict_19.rds", package = "CytoTRACE2"))
predicted_df <- predictData(parameter_dict, ranked_data, log2_data, parallelize_models = TRUE, ncores = 4)
}
