% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessData}
\alias{preprocessData}
\title{Preprocess Data}
\usage{
preprocessData(data, species)
}
\arguments{
\item{data}{The input data to be preprocessed, output of loadData function
(count matrix, samples as columns, genes as rows)}

\item{species}{String, indicates the species name from which the gene names for the input data come from
(2 options: human or mouse)}
}
\value{
Mapped and ranked expression matrix (samples as rows, genes as columns)
}
\description{
This function maps the gene names (rownames) of input data to mouse gene names
based on a provided mapping dictionary, filters and orders the genes based on provided feature set,
and ranks the genes for each sample.
Caution: Please make sure that the input dataset contains gene symbols from only one species (either human or mouse),
following the gene symbol convention from HGNC for human and MGI for mouse.
}
\examples{
ranked_data <- preprocessData(dt)

}
