% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{scale_rows_custom}
\alias{scale_rows_custom}
\title{Scale function
This custom function scales the rows of a matrix to have a mean of 0 and a standard deviation of 1.
It handles cases where the standard deviation of a row is very small, and mean of the scaled row is not close to 0.
It uses population standard deviation instead of sample standard deviation.}
\usage{
scale_rows_custom(
  x,
  constant_threshold = 10 * .Machine$double.eps,
  mean_tolerance = 1e-10
)
}
\arguments{
\item{x}{A matrix of numeric values where rows are cells and columns are genes.}

\item{constant_threshold}{A numeric value representing the threshold below which the standard deviation of a row is considered very small (default is 10 * .Machine$double.eps).}

\item{mean_tolerance}{A numeric value representing the tolerance for the mean of the scaled row to be considered close to 0 (default is 1e-10).}
}
\value{
A matrix of scaled values where rows have a mean of 0 and a standard deviation of 1.
}
\description{
Scale function
This custom function scales the rows of a matrix to have a mean of 0 and a standard deviation of 1.
It handles cases where the standard deviation of a row is very small, and mean of the scaled row is not close to 0.
It uses population standard deviation instead of sample standard deviation.
}
