% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{shortest_consensus}
\alias{shortest_consensus}
\title{Shortest Consensus Function
This function finds the smallest segment size in the predicted potency scores of the neighboring cells
where the average scores of two consecutive portions fall into the same potency category
and returns twice that segment length.}
\usage{
shortest_consensus(neighbor_scores)
}
\arguments{
\item{neighbor_scores}{A numeric vector of predicted potency scores of neighboring cells.}
}
\value{
An integer representing  2*shortest segment size
}
\description{
Shortest Consensus Function
This function finds the smallest segment size in the predicted potency scores of the neighboring cells
where the average scores of two consecutive portions fall into the same potency category
and returns twice that segment length.
}
