% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{smoothData}
\alias{smoothData}
\title{Smoothing}
\usage{
smoothData(
  mat,
  predicted_df,
  top_genes,
  parallelize_smoothing,
  ncores,
  smooth_batch_size,
  seed
)
}
\arguments{
\item{mat}{matrix of ranked gene expression where rows are cells and columns are genes}

\item{predicted_df}{dataframe of 2 columns: predicted potency score 
and predicted potency category for each cell (CytoTRACE2 model predicted values from predict function, prior to postprocessing)}

\item{top_genes}{list of the top 1000 most variable genes (output of dispersion function)}

\item{parallelize_smoothing}{Logical, whether to run the function on subsamples in parallel on multiple threads (default is TRUE).}

\item{ncores}{integer indicating the number of cores to utilize}

\item{smooth_batch_size}{integer indicating the number of cells to subsample}

\item{seed}{Integer, specifying the seed for reproducibility in random processes (default is 14).}
}
\value{
a numeric vector of the predicted potency score of single cells from 1.0 (least differentiated) to 0.0 (most differentiated)
}
\description{
This function generates single-cell predictions of differentiation status from scRNA-seq data.
It takes in a matrix of gene expression values where columns are cells and rows are genes.\cr\cr
annotations matching the number of columns (cells) in the matrix. Furthermore, for the analysis of
large datasets (>3,000 cells), users can increase the speed performance by enabling a subsampling approach for calculation
and using multiple cores.
}
\examples{
mat --> first element in the output of preprocessData()
predicted_df --> output of predictData()
SmoothScore <- smoothData(expression_data, predicted_df, ncores = 4, parallelize_smoothing = TRUE, smooth_batch_size = 1000)
}
