% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{softmax}
\alias{softmax}
\title{Softmax}
\usage{
softmax(x)
}
\arguments{
\item{x}{The input vector.}
}
\value{
The vector after applying softmax.
}
\description{
This function applies the softmax function to the input vector.
}
\examples{
softmax_values <- softmax(x)
}
