% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{smoothDatakNN}
\alias{smoothDatakNN}
\title{Smooth Data with kNN}
\usage{
smoothDatakNN(log2_df, predicted_df, seed, ncores)
}
\arguments{
\item{log2_df}{matrix of log2-transformed gene expression values where rows are cells and columns are genes}

\item{predicted_df}{dataframe of 4 columns containing CytoTRACE 2 model predicted and postprocessed potency scores and categories}

\item{seed}{integer specifying the seed for reproducibility in random processes (default is 14).}
}
\value{
A dataframe with final updated predictions
}
\description{
This applies adaptive nearest neighbor smoothing to datasets with >100 cells.
It identifies the nearest neighbors of each cell based on top principal components,
recalculates each cell’s potency score as a distance-weighted average of its neighbors 
including itself, and updates potency category accordingly.
}
\examples{
 log2_data --> second element in the output of preprocessData()
 predicted_df --> output of binData()
 predicted_df <- smoothDatakNN(ranked_df, predicted_df, seed = 14, ncores = 4)
}
